backend/app/bot.py

    # Placeholder for telegram bot logic. Replace with full code when ready.
    from telegram import Update
    from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes
    from .config import settings

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
        await update.message.reply_text('Bot running')
